code review for Concept Insights v2
* `concept_fields` and `document_fields` are json fields which make it
hard for users to create them. We need to investigate if there is a way
to make them be primitive objects like booleans or lists and create the
json representation during the service call. Another option is to
create POJOs that represent them and use Gson to convert them to Json.
* Javadoc for the params map now includes a <ul> element.
* Added {@link POJO class} when describing the return of a service call.
* Update body parameter to be text for annotatedText.
* Use identifier not “id” as parameter description
* Use `application/json` as MediaType when sending json.
* Switch from `params` to `parameters` as parameter name.
* Switch path constants to be private.
* Updated `toString()` should print the json format.
* Removed arrays initialization in POJOs.