[DCS] Adds required version query param
The Document Conversion Service will soon require a "version" query
parameter on all requests. In the future, this will be used to avoid
breaking changes to behavior when the service is upgraded to newer
versions.

The empty constructor will use a version date set to yesterday's date
(in UTC). However, it should be noted that not specifying a fixed value
may result in future requests failing or outputting unexpected results
due to new or changed behavior or API/behavior-breaking changes as the
service is upgraded. Thus the empty constructor has been deprecated.