[document-conversion] Adds fields to index config
This commit adds the 'fields' object to the IndexConfiguration.
This will allow users to configure how they want to handle fields
when using index_document API.  It will include field mappings,
fields to include, and fields to exclude.

The examples and tests have been updated to make use of this object.