[document-conversion] Refactor IndexFields object
This commit refactors the IndexFields object. It
will now be constructed with a Builder pattern
and allow the user to specify the fields they
want to map, the fields they want to include,
and the fields they want to exclude. The IndexFields
object will can be optionally passed to the
IndexConfiguration.