STT: Close websocket after all data processed
The Speech to Text WebSocketManager never attemtps to close the
websocket used when streaming audio to the server.

According to the API docs, after a request is fully processed, the
server sends a state message, indicating that it is ready to process a
new request. If the thread responsible for sending data to the server
has terminated when we receive such a message, the websocket is now
closed.